+++ src/util.h
@@ -7,6 +7,6 @@ void tacc_die(char *err, ...);
 void tacc_assert(int cond, char *err, ...);
 void *tacc_malloc(size_t sz);
 int tacc_hex_to_dec(char hex);
-int tacc_sizeadj(int count, int size);
+size_t tacc_sizeadj(size_t count, size_t size);
 
 #endif
